Solution to Document
Problem Statement
Problem 1:
Create an HBase table named 'clicks' with a column family 'hits' such that it should be able to
store last 5 values of qualifiers inside 'hits' column family.

Problem 2:
Add few records in the table and update some of them. Use IP Address as row-key. Scan the
table to view if all the previous versions are getting displayed.